feat: Add functionality to jump to last selected worktree
Added the `worktree_go_last` shell function and documented it in the README.
This function allows users to quickly navigate to the most recently selected
worktree within a repository. This is achieved by reading a `.last-selected`
file maintained by `lazyworktree` within the worktree's specific directory. The
README was also updated to reflect new examples for jumping directly to a
worktree and utilizing the `worktree_go_last` alias.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>