feat: Add delta arguments to config
Added a new configuration option `delta_args` to allow users to specify custom
arguments for the `delta` diff tool. This provides greater flexibility in
customizing the appearance and behavior of diffs.

The configuration can accept either a string or a list of strings for
`delta_args`. Default arguments include `--syntax-theme` and `Dracula`. The
`git.Service` was updated to accept and pass these arguments to the `delta`
command. Documentation and example configuration files were also updated to
reflect this change.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>